Add /ignore command that resets on server restart.
This introduces two new commands, `/ignore <username>` and `/unignore <username>` which each add and remove players from a player's ignore list respectively.

If a player is ignoring another, they will not see chat messages or PMs from that player.
The permission 'nerdmessage.ignore.bypass-msg' is intended for staff, and allows them to still send personal messages to a player that is ignoring them.